move poster code into its own file
