+++ cwa_qr/__init__.py
@@ -1,5 +1,6 @@
-from .cwa import (CwaEventDescription, CwaLocation, CwaPoster, generate_payload,
-                  generate_poster, generate_qr_code, generate_url, lowlevel)
+from .cwa import (CwaEventDescription, CwaLocation, generate_payload,
+                  generate_qr_code, generate_url, lowlevel)
+from .poster import CwaPoster, generate_poster
 
 from .rollover import rollover_date
